Move fullscreen hook to main overrides file
